Rearrange downloads code into typical Django modules structure
